refactor: rename configure.h to cose_configure.h
