+++ include/cose/cose_configure.h
