+++ src/Cose.c
@@ -5,7 +5,7 @@
 
 #include "cose/cose.h"
 #include "cose_int.h"
-#include "cose/configure.h"
+#include "cose/cose_configure.h"
 #include "crypto.h"
 
 bool IsValidCOSEHandle(HCOSE h)
