+++ src/Encrypt.c
@@ -13,7 +13,7 @@
 
 #include "cose/cose.h"
 #include "cose_int.h"
-#include "cose/configure.h"
+#include "cose/cose_configure.h"
 #include "crypto.h"
 
 #if INCLUDE_ENCRYPT || INCLUDE_MAC
