+++ src/MacMessage.c
@@ -12,7 +12,7 @@
 
 #include "cose/cose.h"
 #include "cose_int.h"
-#include "cose/configure.h"
+#include "cose/cose_configure.h"
 #include "crypto.h"
 
 #if INCLUDE_MAC
