+++ src/Recipient.c
@@ -5,7 +5,7 @@
 
 #include "cose/cose.h"
 #include "cose_int.h"
-#include "cose/configure.h"
+#include "cose/cose_configure.h"
 #include "crypto.h"
 
 #if INCLUDE_ENCRYPT || INCLUDE_ENCRYPT0 || INCLUDE_MAC || INCLUDE_MAC0
