+++ test/mac_test.c
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <cose/cose.h>
-#include <cose/configure.h>
+#include <cose/cose_configure.h>
 #include <cn-cbor/cn-cbor.h>
 #if INCLUDE_MAC && !INCLUDE_ENCRYPT0
 #include <cose_int.h>
