+++ test/test.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <cose/cose.h>
-#include <cose/configure.h>
+#include <cose/cose_configure.h>
 #include <cn-cbor/cn-cbor.h>
 #include <assert.h>
